Replaces Master-Slave with Client-Server
Expunging all occurrences of inappropriate language of the query and
response paradigm of PLC communications. We are familiar with the Modbus
terminology and support this change.
https://www.modbus.org/docs/Client-ServerPR-07-2020-final.docx.pdf

An adapted version already exists for the server side. Since the server
side cannot be updated immediately, there is still a translation
function in this project to maintain compatibility.